Make a subscribeCustomer() function
I called this an 'action' which may be a little overloaded. Since it doesn't really have any hooks that are dependencies, I wasn't sure it made sense to make _this_ a hook as well.
Why not make the storage thing use hooks, you might ask? Because the calls to 'set' might not happen in the same application scope as the calls to 'get', and because it's probably a bad idea to make application state depend on globals that can change.

So, with all that in mind, I made a function that consumes what's currently on the global state and does API stuff with it. It's not the slickest thing ever, but it seems good and I think it should work (hard to say without some more API stuff and the billing component in place)