friendship over with microbundle, tsc is new best friend
microbundle was causing headaches — too much of an abstraction around rollup. I wanted to use ESBuild, but we need to emit type declaration files and that 'is not something that esbuild can do for you', according to their docs. So, I'm using plain ol' tsc. It works just about as fast and it emits type files. What's not to love.