Remove custom Stripe hook, in-memory store
Having Stripe creds stored in-memory is only useful if we need to share those credentials between apps on the same page (like we might with the state of a signup flow). It looks like the shared-memory storage of the Intent token was only used to facilitate that token between two components, both of which were in PaymentSubmission.tsx. I took that bit out to simplify the code a little. I also removed the custom hook that just hit the API endpoint to get the intent. As written, it didn't have the intended effect of making the call and then re-rendering the component once the call returned (with the assumption that the component would get the updated Intent from the shared memory). I opted to just use a simple useEffect and useState, which seems a little more straightforward to me.