Remove global state from PlanPicker
PlanPicker now shows your plan options and also highlights the plan you have currently, if it's an option for you to subscribe to. It does not update global state, and instead relies on yout o provide a callback, where it provides the selected plan name to you. It's up to you to call subscribeCustomer() with it.