Use all relative imports
Unfortunately, it seems like TS library modules have to have relative imports. This is because when the project is transpiled, the imports are left alone. Unless you do some fanciness with your bundler, this causes types to not be resolved properly when you distribute your package.
Read more on this here:
https://stackoverflow.com/questions/56949487/how-to-have-absolute-import-paths-in-a-library-project