Add a workflow to publish to NPM
This workflow publishes this package to NPM. Note that I'm using NPM here, rather than Yarn, on the off-chance the npm-publish workflow requires an NPM-install first.

This workflow config is identical to the one in the [npm-publish](https://github.com/marketplace/actions/npm-publish) example with the exception of the additional `npm run build` step (we distribute the `/dist` directory, which needs to be built). Technically speaking `npm test` should also run a build, but better to be safe (explicit) than sorry (accidentally distributing an un-built project).